requirements.txt
fastapi
uvicorn[standard]
sqlmodel
email-validator
psycopg2-binary
